requirements.txt
passlib==1.7.1
texttable==1.6.1
PyYAML==5.3.1
